Integrate image parser with the main parser
